refactor: move util from steps to root folder
This is done because the helper functions will be needed by code outside the
steps.
